Update request.php
Using a login system for this file is really suggested because this file is the one that handle the database requests. If you do not know how to create a login system you can found one here: http://www.php-login.net/

The minimal version with some changes is enough to keep your database secure.